Move ohif:hotkeys into ohif:core
