+++ Packages/ohif-core/client/components/hotkeys/formTable.html
