+++ Packages/ohif-core/client/hotkeys.js
@@ -1,5 +1,5 @@
 import { OHIF } from 'meteor/ohif:core';
-import { HotkeysManager } from 'meteor/ohif:hotkeys/client/classes/HotkeysManager';
+import { HotkeysManager } from '/client/classes/HotkeysManager';
 import 'jquery.hotkeys';
 
 // Create hotkeys namespace using a HotkeysManager class instance
