+++ Packages/ohif-core/client/routes.js
@@ -12,3 +12,16 @@ if (Meteor.isDevelopment) {
         this.render('componentPlayground');
     });
 }
+
+import { hotkeys } from 'meteor/ohif:hotkeys';
+
+Router.onBeforeAction(function() {
+    const lastRoute = Session.get('lastRoute');
+    const currentRoute = this.router.current().route.getName();
+    if (currentRoute !== lastRoute) {
+        hotkeys.switchToContext(null);
+        Session.set('lastRoute', currentRoute);
+    }
+
+    this.next();
+});
