--- Packages/ohif-hotkeys/client/index.js
@@ -1,2 +1,0 @@
-import './components';
-import './lib';
