--- Packages/ohif-hotkeys/client/lib/index.js
@@ -1,1 +1,0 @@
-import './routes';
